+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added a podcast general settings page
 - Added an option to disallow publishing RSS for a podcast
+- Each podcast has two settings pages (general settings, episode settings)
 - Fixed a bug where changing lightswitch inputs on the sidebar of the episode edit page, creates unwanted drafts
 
 ## 0.1.0 - 2023-03-02